requirements.txt
beautifulsoup4==4.14.3
langdetect==1.0.9
pycountry==24.6.1
python-dotenv==1.2.1
requests==2.32.5
spotipy==2.25.2
